third_party/WebKit/LayoutTests/inspector-protocol/css/css-get-media-queries.html

<html>
<head>
<script type="text/javascript" src="../../http/tests/inspector-protocol/inspector-protocol-test.js"></script>
<script type="text/javascript">
function test()
{
    InspectorTest.sendCommandOrDie("DOM.enable", {});
    InspectorTest.sendCommandOrDie("CSS.enable", {}, onCSSEnabled);

    function onCSSEnabled()
    {
        InspectorTest.sendCommandOrDie("CSS.getMediaQueries", {}, onMediaQueries);
    }

    function onMediaQueries(response)
    {
        function compareValue(value1, value2)
        {
            return value1 < value2 ? -1 : value1 > value2 ? 1 : 0;
        }

        function mediaComparator(media1, media2)
        {
            if (media1.text != media2.text)
                return compareValue(media1.text, media2.text);
            var hasSource1 = !!(media1.styleSheetId && media1.range);
            var hasSource2 = !!(media2.styleSheetId && media2.range);
            if (hasSource1 != hasSource2)
                return compareValue(hasSource1, hasSource2);
            if (media1.range.startLine != media2.range.startLine)
                return compareValue(media1.range.startLine, media2.range.startLine);
            if (media1.range.startColumn != media2.range.startColumn)
                return compareValue(media1.range.startColumn, media2.range.startColumn);
            if (media1.range.endLine != media2.range.endLine)
                return compareValue(media1.range.endLine, media2.range.endLine);
            if (media1.range.endColumn != media2.range.endColumn)
                return compareValue(media1.range.endColumn, media2.range.endColumn);
            return 0;
        }
        response.medias.sort(mediaComparator);

        var styleSheetIds = [];
        for (var i = 0; i < response.medias.length; ++i) {
            var styleSheetId = response.medias[i].styleSheetId;
            if (styleSheetId && styleSheetIds.indexOf(styleSheetId) === -1)
                styleSheetIds.push(styleSheetId);
        }
        if (!styleSheetIds.length) {
            dumpMediaQueries(response.medias, {});
            return;
        }

        var styleSheetTextMap = {};
        for (var i = 0; i < styleSheetIds.length; ++i) {
            var callback = i === styleSheetIds.length - 1 ? dumpMediaQueries.bind(null, response.medias, styleSheetTextMap) : null;
            InspectorTest.sendCommandOrDie("CSS.getStyleSheetText", {styleSheetId: styleSheetIds[i]}, onStyleSheetText.bind(null, styleSheetIds[i], styleSheetTextMap, callback));
        }
    }

    function onStyleSheetText(styleSheetId, styleSheetTextMap, callback, response)
    {
        styleSheetTextMap[styleSheetId] = response.text;
        if (callback)
            callback();
    }

    function dumpMediaQueries(medias, styleSheetTextMap)
    {
        for (var i = 0; i < medias.length; ++i) {
            var mediaRule = medias[i];
            InspectorTest.log("mediaRule #" + i);
            InspectorTest.log("    text: " + mediaRule.text);
            InspectorTest.log("    source: " + mediaRule.source);
            InspectorTest.log("    range: " + JSON.stringify(mediaRule.range));
            if (mediaRule.styleSheetId && mediaRule.range)
                InspectorTest.log("    computedText: " + getComputedText(styleSheetTextMap[mediaRule.styleSheetId], mediaRule.range));
            if (!mediaRule.mediaList) {
                InspectorTest.log("    mediaList is empty");
                continue;
            }
            for (var j = 0; j < mediaRule.mediaList.length; ++j) {
                var mediaQuery = mediaRule.mediaList[j];
                var suffix = mediaRule.sourceURL.indexOf("active-media-queries.css") !== -1 ? " active: " + mediaQuery.active : "";
                InspectorTest.log("    mediaQuery #" + j + suffix);
                for (var k = 0; k < mediaQuery.expressions.length; ++k) {
                    var expression = mediaQuery.expressions[k];
                    InspectorTest.log("        mediaExpression #" + k);
                    InspectorTest.log("            feature: " + expression.feature);
                    InspectorTest.log("            value: " + expression.value);
                    InspectorTest.log("            unit: " + expression.unit);
                    if (expression.valueRange) {
                        InspectorTest.log("            valueRange: " + JSON.stringify(expression.valueRange));
                        if (mediaRule.styleSheetId)
                            InspectorTest.log("            computedText: " + getComputedText(styleSheetTextMap[mediaRule.styleSheetId], expression.valueRange));
                    }
                    if (expression.computedLength)
                        InspectorTest.log("            computed length: " + expression.computedLength);
                }
            }
        }
        InspectorTest.completeTest();
    }

    function getComputedText(text, range)
    {
        var lines = text.split("\n");
        var result = [];
        for (var line = range.startLine; line <= range.endLine; ++line) {
            var start = line === range.startLine ? range.startColumn : 0;
            var end = line === range.endLine ? range.endColumn : lines[line].length;
            result.push(lines[line].substring(start, end));
        }
        return result.join("\\n");
    }
};
</script>
<link rel="stylesheet" media="print and (min-width: 8.5in)" href="./resources/media-queries.css"></link>
<link rel="stylesheet" href="./resources/active-media-queries.css"></link>
<style>
@media screen and (device-aspect-ratio: 16/9), screen and (device-aspect-ratio: 16/10) {
    * {
        padding: 1em;
    }

    @media (max-width: 200px) and (min-width: 100px) {
        * {
            margin: 1in;
        }
    }
}
</style>
</head>
<body onload="runTest()">
</body>
</html>
